Reorganize Mappable classes
