+++ app/controllers/locations/maps_controller.rb
@@ -15,7 +15,7 @@ module Locations
       columns = %w[name north south east west].map { |x| "locations.#{x}" }
       args = { select: "DISTINCT(locations.id), #{columns.join(", ")}" }
       @locations = @query.select_rows(args).map do |id, *the_rest|
-        MinimalMapLocation.new(id, *the_rest)
+        Mappable::MinimalLocation.new(id, *the_rest)
       end
     end
   end
